Move agreement-process tests to integration dir
