+++ packages/agreement-process/test/integration/deleteAgreement.test.ts
@@ -19,23 +19,23 @@ import {
   TenantId,
 } from "pagopa-interop-models";
 import { describe, expect, it, vi } from "vitest";
-import { agreementDeletableStates } from "../src/model/domain/agreement-validators.js";
+import { agreementDeletableStates } from "../../src/model/domain/agreement-validators.js";
 import {
   agreementNotFound,
   agreementNotInExpectedState,
   organizationIsNotTheConsumer,
   organizationIsNotTheDelegateConsumer,
-} from "../src/model/domain/errors.js";
-import { config } from "../src/config/config.js";
+} from "../../src/model/domain/errors.js";
+import { config } from "../../src/config/config.js";
 import {
   addOneAgreement,
   addOneDelegation,
   agreementService,
   fileManager,
-  getMockConsumerDocument,
   readLastAgreementEvent,
   uploadDocument,
-} from "./utils.js";
+} from "../integrationUtils.js";
+import { getMockConsumerDocument } from "../mockUtils.js";
 
 describe("delete agreement", () => {
   it.each(agreementDeletableStates)(
