+++ packages/agreement-process/test/integration/internalArchiveAgreementAfterDelegationRevocation.test.ts
@@ -19,17 +19,17 @@ import {
   toAgreementV2,
 } from "pagopa-interop-models";
 import { describe, expect, it, vi } from "vitest";
-import { agreementArchivableStates } from "../src/model/domain/agreement-validators.js";
+import { agreementArchivableStates } from "../../src/model/domain/agreement-validators.js";
 import {
   agreementNotFound,
   agreementNotInExpectedState,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   addOneAgreement,
   addOneDelegation,
   agreementService,
   readLastAgreementEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
 describe("internal archive agreement", () => {
   it("should succeed when the agreement is in an archivable state", async () => {
