+++ packages/agreement-process/test/mockUtils.ts
@@ -2,180 +2,31 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { randomInt } from "crypto";
 import {
-  StoredEvent,
-  readLastEventByStreamId,
-  setupTestContainersVitest,
-  writeInEventstore,
-  writeInReadmodel,
-  ReadEvent,
-  readEventByStreamIdAndVersion,
   randomArrayItem,
   getMockDelegation,
   getMockTenant,
   getMockAuthData,
 } from "pagopa-interop-commons-test";
-import { afterAll, afterEach, expect, inject, vi } from "vitest";
 import {
   Agreement,
-  AgreementEvent,
   AgreementId,
-  EService,
   Tenant,
-  toAgreementV2,
-  toReadModelEService,
-  toReadModelTenant,
-  toReadModelAgreement,
   AgreementDocumentId,
   generateId,
   AgreementDocument,
-  Attribute,
-  toReadModelAttribute,
   TenantId,
   Delegation,
   AgreementStamp,
   UserId,
   delegationKind,
   delegationState,
-  ListResult,
 } from "pagopa-interop-models";
 import { agreementApi } from "pagopa-interop-api-clients";
-import {
-  UIAuthData,
-  formatDateyyyyMMddHHmmss,
-  genericLogger,
-  initPDFGenerator,
-  launchPuppeteerBrowser,
-} from "pagopa-interop-commons";
-import puppeteer, { Browser } from "puppeteer";
+import { UIAuthData, formatDateyyyyMMddHHmmss } from "pagopa-interop-commons";
 import { subDays } from "date-fns";
 import { match } from "ts-pattern";
 import { z } from "zod";
-import { agreementServiceBuilder } from "../src/services/agreementService.js";
-import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { config } from "../src/config/config.js";
-import { contractBuilder } from "../src/services/agreementContractBuilder.js";
-
-export const { cleanup, readModelRepository, postgresDB, fileManager } =
-  await setupTestContainersVitest(
-    inject("readModelConfig"),
-    inject("eventStoreConfig"),
-    inject("fileManagerConfig")
-  );
-
-afterEach(cleanup);
-
-const testBrowserInstance: Browser = await launchPuppeteerBrowser({
-  pipe: true,
-});
-const closeTestBrowserInstance = async (): Promise<void> =>
-  await testBrowserInstance.close();
-
-afterAll(closeTestBrowserInstance);
-
-vi.spyOn(puppeteer, "launch").mockImplementation(
-  async () => testBrowserInstance
-);
-
-export const { agreements, attributes, eservices, tenants, delegations } =
-  readModelRepository;
-
-export const readModelService = readModelServiceBuilder(readModelRepository);
-
-export const pdfGenerator = await initPDFGenerator();
-
-export const agreementContractBuilder = contractBuilder(
-  readModelService,
-  pdfGenerator,
-  fileManager,
-  config,
-  genericLogger
-);
-
-export const agreementService = agreementServiceBuilder(
-  postgresDB,
-  readModelService,
-  fileManager,
-  pdfGenerator
-);
-export const writeAgreementInEventstore = async (
-  agreement: Agreement
-): Promise<void> => {
-  const agreementEvent: AgreementEvent = {
-    type: "AgreementAdded",
-    event_version: 2,
-    data: { agreement: toAgreementV2(agreement) },
-  };
-  const eventToWrite: StoredEvent<AgreementEvent> = {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    stream_id: agreementEvent.data.agreement!.id,
-    version: 0,
-    event: agreementEvent,
-  };
-
-  await writeInEventstore(eventToWrite, "agreement", postgresDB);
-};
-
-export const addOneAgreement = async (agreement: Agreement): Promise<void> => {
-  await writeAgreementInEventstore(agreement);
-  await writeInReadmodel(toReadModelAgreement(agreement), agreements);
-};
-
-export const addOneEService = async (eservice: EService): Promise<void> => {
-  await writeInReadmodel(toReadModelEService(eservice), eservices);
-};
-
-export const addOneTenant = async (tenant: Tenant): Promise<void> => {
-  await writeInReadmodel(toReadModelTenant(tenant), tenants);
-};
-
-export const addOneAttribute = async (attribute: Attribute): Promise<void> => {
-  await writeInReadmodel(toReadModelAttribute(attribute), attributes);
-};
-
-export const addOneDelegation = async (
-  delegation: Delegation
-): Promise<void> => {
-  await writeInReadmodel(delegation, delegations);
-};
-
-export const readLastAgreementEvent = async (
-  agreementId: AgreementId
-): Promise<ReadEvent<AgreementEvent>> =>
-  await readLastEventByStreamId(agreementId, "agreement", postgresDB);
-
-export const readAgreementEventByVersion = async (
-  agreementId: AgreementId,
-  version: number
-): Promise<ReadEvent<AgreementEvent>> =>
-  await readEventByStreamIdAndVersion(
-    agreementId,
-    version,
-    "agreement",
-    postgresDB
-  );
-
-export async function uploadDocument(
-  agreementId: AgreementId,
-  documentId: AgreementDocumentId,
-  name: string
-): Promise<void> {
-  const documentDestinationPath = `${config.consumerDocumentsPath}/${agreementId}`;
-  await fileManager.storeBytes(
-    {
-      bucket: config.s3Bucket,
-      path: documentDestinationPath,
-      resourceId: documentId,
-      name,
-      content: Buffer.from("large-document-file"),
-    },
-    genericLogger
-  );
-  expect(
-    await fileManager.listFiles(config.s3Bucket, genericLogger)
-  ).toContainEqual(
-    `${config.consumerDocumentsPath}/${agreementId}/${documentId}/${name}`
-  );
-}
 
 export function getMockConsumerDocument(
   agreementId: AgreementId,
@@ -350,36 +201,3 @@ export const authDataAndDelegationsFromRequesterIs = (
       };
     })
     .exhaustive();
-
-export async function addDelegationsAndDelegates({
-  producerDelegation,
-  delegateProducer,
-  consumerDelegation,
-  delegateConsumer,
-}: {
-  producerDelegation: Delegation | undefined;
-  delegateProducer: Tenant | undefined;
-  consumerDelegation: Delegation | undefined;
-  delegateConsumer: Tenant | undefined;
-}): Promise<void> {
-  if (producerDelegation && delegateProducer) {
-    await addOneDelegation(producerDelegation);
-    await addOneTenant(delegateProducer);
-  }
-
-  if (consumerDelegation && delegateConsumer) {
-    await addOneDelegation(consumerDelegation);
-    await addOneTenant(delegateConsumer);
-  }
-}
-
-export function expectSinglePageListResult<T>(
-  actual: ListResult<T>,
-  expected: T[]
-): void {
-  expect(actual).toEqual({
-    totalCount: expected.length,
-    results: expect.arrayContaining(expected),
-  });
-  expect(actual.results).toHaveLength(expected.length);
-}
